+++ ReleaseNotes.md
@@ -21,4 +21,7 @@ shows an error dialog
 * **feature:** Authentication for Discogs now with OAuth access so that no personal user token of the user is required
 * **feature:** Genrers can be managed now in a dialog: change names, create new genres and delete unused genres
 * **feature:** Interprets can be managed now in a dialog: change names, create new interprets and delete unused interprets
+* **feature:** Size of thumbnails can now be sized with a slider at the top of the thumbnail panel. The size is saved and restored at the next start of the app.
+* **feature:** Advanced search dialog for albums (via menu, button over thumbnail panel or hot key CTRL-F) 
+